+++ tests/accelerators/test_accelerator_connector.py
@@ -28,7 +28,7 @@ from pytorch_lightning.plugins import (
     DataParallelPlugin,
     DDP2Plugin,
     DDPPlugin,
-    DDPShardedPlugin,
+    DDPShardedStrategy,
     DDPSpawnPlugin,
     DDPSpawnShardedPlugin,
     DeepSpeedPlugin,
@@ -406,11 +406,11 @@ def test_plugin_accelerator_choice(accelerator: Optional[str], plugin: str):
     else:
         with pytest.deprecated_call(match=r"accelerator=.*\)` has been deprecated"):
             trainer = Trainer(accelerator=accelerator, plugins=plugin, num_processes=2)
-    assert isinstance(trainer.training_type_plugin, DDPShardedPlugin)
+    assert isinstance(trainer.training_type_plugin, DDPShardedStrategy)
 
     with pytest.deprecated_call(match="Passing .* `strategy` to the `plugins`"):
         trainer = Trainer(plugins=plugin, num_processes=2)
-    assert isinstance(trainer.training_type_plugin, DDPShardedPlugin)
+    assert isinstance(trainer.training_type_plugin, DDPShardedStrategy)
 
 
 @pytest.mark.parametrize(
@@ -418,7 +418,7 @@ def test_plugin_accelerator_choice(accelerator: Optional[str], plugin: str):
     [
         ("ddp", DDPPlugin),
         ("ddp_spawn", DDPSpawnPlugin),
-        ("ddp_sharded", DDPShardedPlugin),
+        ("ddp_sharded", DDPShardedStrategy),
         ("ddp_sharded_spawn", DDPSpawnShardedPlugin),
         pytest.param("deepspeed", DeepSpeedPlugin, marks=RunIf(deepspeed=True)),
     ],
@@ -629,7 +629,7 @@ def test_strategy_choice_cpu_plugin(tmpdir, plugin):
         ("ddp_find_unused_parameters_false", DDPPlugin),
         ("ddp2", DDP2Plugin),
         ("dp", DataParallelPlugin),
-        ("ddp_sharded", DDPShardedPlugin),
+        ("ddp_sharded", DDPShardedStrategy),
         ("ddp_sharded_spawn", DDPSpawnShardedPlugin),
         pytest.param("deepspeed", DeepSpeedPlugin, marks=RunIf(deepspeed=True)),
     ],